Insert und update Method hinzufügen

+++ src/main/java/iuinformatik/busreisen/busreisen_oop2_referat/database/DBType.java
@@ -5,7 +5,7 @@ public enum DBType {
     LongString("VARCHAR(255)"),
     Bool("BOOL"),
     Int("INT"),
-    Float("FLOAT"),
+    Double("NUMERIC"),
     Date("DATE"),
     Time("TIME"),
     TimeStamp("TIMESTAMP");
